+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file tracks changes to this project. It follows the [Keep a Changelog format](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.5.2] - 2023-07-12
 
 ### Added
 
@@ -147,6 +147,8 @@ Documentation and examples were also fixed and improved.
 [#5]: https://github.com/tmaxmax/go-sse/pull/5
 [#2]: https://github.com/tmaxmax/go-sse/pull/2
 
+[0.5.2]: https://github.com/tmaxmax/go-sse/releases/tag/v0.5.2
+[0.5.1]: https://github.com/tmaxmax/go-sse/releases/tag/v0.5.1
 [0.5.0]: https://github.com/tmaxmax/go-sse/releases/tag/v0.5.0
 [0.4.3]: https://github.com/tmaxmax/go-sse/releases/tag/v0.4.3
 [0.4.2]: https://github.com/tmaxmax/go-sse/releases/tag/v0.4.2